requirements.txt
azure_storage==0.37.0
Flask==3.1.0
msal==1.31.1
openpyxl==3.1.5
pandas==2.2.3
Requests==2.32.3
azure-storage-file-datalake
azure.identity